# core: Overwrite values for existing custom call option keys, instead of keeping old values around.