cc: @dapengzhang0 

This PR resulted from this discussion https://github.com/grpc/grpc-java/pull/1918#discussion_r66713223

The performance of this PR seems identical with the current master and it fixes `toString` when having overwritten existing keys. Another upside of this change is that there is no unexpected performance degradation in case someone frequently changes the value of the same key.

**This PR:**
```
CallOptionsBenchmark.withOption                               1  sample  184633    253.033 ±  3.066  ns/op
CallOptionsBenchmark.withOption                               2  sample  163158    528.771 ± 14.739  ns/op
CallOptionsBenchmark.withOption                               4  sample  129367   1277.749 ± 60.131  ns/op
CallOptionsBenchmark.withOption                               8  sample  185144   3421.190 ± 23.552  ns/op
CallOptionsBenchmark.withOption                              16  sample  121435  10338.319 ± 50.105  ns/op
CallOptionsBenchmark.withOptionDuplicates                     1  sample  161893     59.429 ±  1.184  ns/op
CallOptionsBenchmark.withOptionDuplicates                     2  sample  142616    316.315 ±  1.904  ns/op
CallOptionsBenchmark.withOptionDuplicates                     4  sample  136853   1197.563 ± 26.170  ns/op
CallOptionsBenchmark.withOptionDuplicates                     8  sample  145914   4336.151 ± 21.998  ns/op
CallOptionsBenchmark.withOptionDuplicates                    16  sample  164104  15307.636 ± 68.240  ns/op
```

**Current master:**
```
Benchmark                                  (customOptionsCount)    Mode     Cnt      Score     Error  Units
CallOptionsBenchmark.withOption                               1  sample  185854    248.739 ±   1.742  ns/op
CallOptionsBenchmark.withOption                               2  sample  160535    525.485 ±   2.570  ns/op
CallOptionsBenchmark.withOption                               4  sample  130372   1243.345 ±  12.546  ns/op
CallOptionsBenchmark.withOption                               8  sample  188768   3371.729 ±  33.134  ns/op
CallOptionsBenchmark.withOption                              16  sample  122475  10260.304 ±  54.070  ns/op
CallOptionsBenchmark.withOptionDuplicates                     1  sample  161350     68.328 ±  13.531  ns/op
CallOptionsBenchmark.withOptionDuplicates                     2  sample  123365    359.238 ±  11.963  ns/op
CallOptionsBenchmark.withOptionDuplicates                     4  sample  106779   1518.031 ±  26.066  ns/op
CallOptionsBenchmark.withOptionDuplicates                     8  sample  110973   5675.588 ±  37.980  ns/op
CallOptionsBenchmark.withOptionDuplicates                    16  sample  112241  22359.046 ± 116.701  ns/op
```