@buchgr 2016-06-13
@carl-mastrangelo I updated the code as you suggested. I had already tested it before and it's exactly the same speed. I assume you prefer this style? I don't have a preference.